added support for altitude
added support for altitude

+++ Inc/hts221.h
@@ -23,8 +23,8 @@ uint8_t hts221_read_byte(uint8_t reg_addr);
 void hts221_readArray(uint8_t * data, uint8_t reg, uint8_t length);
 uint8_t hts221_init(void);
 void hts221_write_byte(uint8_t reg_addr, uint8_t value);
-int16_t hts221_get_temperature();
-uint16_t hts221_get_humidity();
+float hts221_get_temperature();
+float hts221_get_humidity();
 
 #endif /* HTS221_H_ */
